refactor/files: delete unwanted files
